+++ Game.Server/Models/Resources/Resource.cs
@@ -13,6 +13,6 @@
 
         public string Name { get; internal set; }
 
-        public int Value { get; internal set; }
+        public float Value { get; internal set; }
     }
 }
